Use key_rev_value with latest_hash policy for latest storage
