+++ test/features/buckets/revisioned_bucket.js
@@ -195,6 +195,9 @@ describe('Revisioned buckets', function() {
             .then(function(res) {
                 assert.deepEqual(res.body.toString(), 'Newer_Revision');
                 assert.deepEqual(res.headers.etag, mwUtils.makeETag(1001, newerUUID));
+            })
+            .catch(function(e) {
+                console.log(e);
             });
         });
 
@@ -234,5 +237,5 @@ describe('Revisioned buckets', function() {
 
     describe('key_rev_value', function() { runTests('key_rev_value') });
     describe('key_rev_large_value', function() { runTests('key_rev_large_value'); });
-    describe('key_rev_archival_value', function() { runTests('key_rev_archival_value'); });
+    describe('key_rev_latest_value', function() { runTests('key_rev_latest_value'); });
 });
